% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_BCa_nt1BR.R
\docType{data}
\name{ind_BCa_nt1BR}
\alias{ind_BCa_nt1BR}
\title{ind_BCa_nt1BR}
\format{
Logical vector of length 60.
}
\usage{
ind_BCa_nt1BR
}
\description{
Variable selection results for the vignette.
}
\examples{

data(ind_BCa_nt1BR)
## maybe str(ind_BCa_nt1BR) ; plot(ind_BCa_nt1BR) ...

}
\references{
Régression Bêta PLS. (French) [PLS Beta regression.], 
F. Bertrand, N. Meyer, M. Beau-Faller, K. El Bayed, N. Izzie-J., 
M. Maumy-Bertrand, (2013), J. SFdS, 154(3):143-159 \cr

Partial Least Squares Regression for Beta Regression Models. 
F. Bertrand, M. Maumy (2021). useR! 2021, Zurich.\cr
}
\keyword{datasets}
