% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modpls.boot3.R
\docType{data}
\name{modpls.boot3}
\alias{modpls.boot3}
\title{Bootstrap distribution of a 3 components model}
\format{
a class boot object
}
\usage{
modpls.boot3
}
\description{
A precomputed bootstrap distribution of the coefficients of a model used in the vignette.
}
\examples{

data(modpls.boot3)
str(modpls.boot3)
plot(modpls.boot3)

}
\references{
Régression Bêta PLS. (French) [PLS Beta regression.], 
F. Bertrand, N. Meyer, M. Beau-Faller, K. El Bayed, N. Izzie-J., 
M. Maumy-Bertrand, (2013), J. SFdS, 154(3):143-159 \cr

Partial Least Squares Regression for Beta Regression Models. 
F. Bertrand, M. Maumy (2021). useR! 2021, Zurich.\cr
}
\keyword{datasets}
