% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modpls_sub4.R
\docType{data}
\name{modpls_sub4}
\alias{modpls_sub4}
\title{A plsRbetamodel model on a data subset}
\format{
a class plsRbetamodel object
}
\usage{
modpls_sub4
}
\description{
A precomputed four components plsRbetamodel fitted to a subset of an example dataset and used in the vignette.
}
\examples{

data(modpls_sub4)
str(modpls_sub4)

}
\references{
Régression Bêta PLS. (French) [PLS Beta regression.], 
F. Bertrand, N. Meyer, M. Beau-Faller, K. El Bayed, N. Izzie-J., 
M. Maumy-Bertrand, (2013), J. SFdS, 154(3):143-159 \cr

Partial Least Squares Regression for Beta Regression Models. 
F. Bertrand, M. Maumy (2021). useR! 2021, Zurich.\cr
}
\keyword{datasets}
