% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TxTum.mod.bootBC1.R
\docType{data}
\name{TxTum.mod.bootBC1}
\alias{TxTum.mod.bootBC1}
\title{Bootstrap distribution TxTum BC1 model}
\format{
a class boot object
}
\usage{
TxTum.mod.bootBC1
}
\description{
A precomputed bootstrap distribution of the coefficients of a model used in the vignette.
}
\examples{

data(TxTum.mod.bootBC1)
str(TxTum.mod.bootBC1)
plot(TxTum.mod.bootBC1)

}
\references{
Régression Bêta PLS. (French) [PLS Beta regression.], 
F. Bertrand, N. Meyer, M. Beau-Faller, K. El Bayed, N. Izzie-J., 
M. Maumy-Bertrand, (2013), J. SFdS, 154(3):143-159 \cr

Partial Least Squares Regression for Beta Regression Models. 
F. Bertrand, M. Maumy (2021). useR! 2021, Zurich.\cr
}
\keyword{datasets}
