% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plsRbeta-package.R
\docType{package}
\name{plsRbeta-package}
\alias{plsRbeta-package}
\title{plsRbeta: Partial Least Squares Regression for Beta Regression Models}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides Partial least squares Regression for (weighted) beta regression models (Bertrand 2013, \url{https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215}) and k-fold cross-validation of such models using various criteria. It allows for missing data in the explanatory variables. Bootstrap confidence intervals constructions are also available.
}
\references{
Partial least squares Regression for (weighted) beta regression models (Bertrand 2013,  <https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215>), \url{https://github.com/fbertran/plsRbeta/} et \url{https://fbertran.github.io/plsRbeta/}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://fbertran.github.io/plsRbeta/}
  \item \url{https://github.com/fbertran/plsRbeta/}
  \item Report bugs at \url{https://github.com/fbertran/plsRbeta/issues/}
}

}
\author{
\strong{Maintainer}: Frederic Bertrand \email{frederic.bertrand@lecnam.net} (\href{https://orcid.org/0000-0002-0837-8281}{ORCID})

Authors:
\itemize{
  \item Myriam Maumy-Bertrand \email{myriam.maumy@ehesp.fr} (\href{https://orcid.org/0000-0002-4615-1512}{ORCID})
}

}
\keyword{internal}
