% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.plsRbetamodel.R
\name{print.plsRbetamodel}
\alias{print.plsRbetamodel}
\title{Print method for plsRbeta models}
\usage{
\method{print}{plsRbetamodel}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{"plsRbetamodel"}}

\item{\dots}{not used}
}
\value{
\code{NULL}
}
\description{
This function provides a print method for the class \code{"plsRbetamodel"}
}
\examples{

data("GasolineYield",package="betareg")
modpls <- plsRbeta(yield~.,data=GasolineYield,nt=3,modele="pls-beta")
print(modpls)

}
\references{
Frédéric Bertrand, Nicolas Meyer,
Michèle Beau-Faller, Karim El Bayed, Izzie-Jacques Namer,
Myriam Maumy-Bertrand (2013). Régression Bêta
PLS. \emph{Journal de la Société Française de Statistique},
\bold{154}(3):143-159.
\url{http://publications-sfds.math.cnrs.fr/index.php/J-SFdS/article/view/215}
}
\seealso{
\code{\link{print}}
}
\author{
Frédéric Bertrand\cr
\email{frederic.bertrand@utt.fr}\cr
\url{https://fbertran.github.io/homepage/}
}
\keyword{methods}
\keyword{print}
