% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.plsRbetamodel.R
\name{print.summary.plsRbetamodel}
\alias{print.summary.plsRbetamodel}
\title{Print method for summaries of plsRbeta models}
\usage{
\method{print}{summary.plsRbetamodel}(x, ...)
}
\arguments{
\item{x}{an object of the class \code{"summary.plsRbetamodel"}}

\item{\dots}{not used}
}
\value{
\item{language}{call of the model}
}
\description{
This function provides a print method for the class
\code{"summary.plsRbetamodel"}
}
\examples{

data("GasolineYield",package="betareg")
modpls <- plsRbeta(yield~.,data=GasolineYield,nt=3,modele="pls-beta")
print(summary(modpls))

}
\references{
Frédéric Bertrand, Nicolas Meyer,
Michèle Beau-Faller, Karim El Bayed, Izzie-Jacques Namer,
Myriam Maumy-Bertrand (2013). Régression Bêta
PLS. \emph{Journal de la Société Française de Statistique},
\bold{154}(3):143-159.
\url{https://ojs-test.apps.ocp.math.cnrs.fr/index.php/J-SFdS/article/view/215}
}
\seealso{
\code{\link{print}} and \code{\link{summary}}
}
\author{
Frédéric Bertrand\cr
\email{frederic.bertrand@lecnam.net}\cr
\url{https://fbertran.github.io/homepage/}
}
\keyword{methods}
\keyword{print}
